Unidad 2/Practicas/requirements.txt
#Miranda Leyva Juan Manuel

mysql-connector-python  #pip install -r requeriments.txt

